refactor(core): decompose bridge.rs into focused submodules [OPE-376]
Split the monolithic GatewayBridge (599 LOC) into focused modules:
- mod.rs: struct definition, construction, lifecycle, accessors
- pairing.rs: pairing code generation and team command dispatch
- relay.rs: message relay and stream driving
- outgoing.rs: outgoing message persistence and event emission
- tests.rs: comprehensive test suite (9 tests, all passing)
